publish zoo wide comments to event stream as well (#110)
* publish zoo wide comments to the stream

* consistent attribute testing for section

check it’s a project not if it’s zooniverse section

* stub the desired section behaviour

deal with stupid call back data setting state based on associations instead of direct attributes being set on create / new, etc. This is one of the reasons i dislike model callbacks so much, it’s hard to understand this without dealing in the murky realm of callbacks.