serialise project title as embedded attribute
add aliased title -> display_name attribute to project as we don’t have access to the contents model via FDW table. The project display name is analogous to the content title we can proxy this attribute and embed the title via the `embed_attributes_from :discussion, :board, :project` in the comment serialiser.